+++ Runtime/Scripts/Export/ImageExportBase.cs
@@ -32,7 +32,9 @@ namespace GLTFast.Export {
         
         public abstract string fileName { get; }
         public abstract string mimeType { get; }
-
+        public abstract FilterMode filterMode { get; }
+        public abstract TextureWrapMode wrapMode { get; }
+        
         public abstract void Write(string filePath, bool overwrite);
         public abstract byte[] GetData();